+++ src/net/player.rs
@@ -6,8 +6,8 @@ pub struct Player {
   pub id: String,
   pub area_id: String,
   pub avatar_id: u16,
-  pub x: f64,
-  pub y: f64,
-  pub z: f64,
+  pub x: f32,
+  pub y: f32,
+  pub z: f32,
   pub ready: bool,
 }
