+++ src/packets/mod.rs
@@ -14,5 +14,5 @@ pub struct PacketHeaders {
   pub id: u64,
 }
 
-pub(super) const TILE_WIDTH: f64 = 62.0 + 2.5;
-pub(super) const TILE_HEIGHT: f64 = 32.0 + 0.5;
+pub(super) const TILE_WIDTH: f32 = 62.0 + 2.5;
+pub(super) const TILE_HEIGHT: f32 = 32.0 + 0.5;
