+++ src/plugins/lua/api/bot_api.rs
@@ -28,7 +28,7 @@ pub fn add_bot_api<'a, 'b>(
   api_table.set(
     "create_bot",
     scope.create_function(
-      move |_, (id, area_id, avatar_id, x, y, z): (String, String, u16, f64, f64, f64)| {
+      move |_, (id, area_id, avatar_id, x, y, z): (String, String, u16, f32, f32, f32)| {
         let mut net = net_ref.borrow_mut();
 
         if let Some(_) = net.get_area(&area_id) {
@@ -108,7 +108,7 @@ pub fn add_bot_api<'a, 'b>(
 
   api_table.set(
     "move_bot",
-    scope.create_function(move |_, (id, x, y, z): (String, f64, f64, f64)| {
+    scope.create_function(move |_, (id, x, y, z): (String, f32, f32, f32)| {
       let mut net = net_ref.borrow_mut();
 
       net.move_bot(&id, x, y, z);
